Refactor step template editor routing and types
Removed legacy ConfigureStepTemplate components and consolidated step template editing to use EDIT_STEP_TEMPLATE route. Introduced a new step-editor-types file for shared StepEditorProps type, updating all relevant imports. Cleaned up main router and related references to streamline workflow editor step template handling.
