+++ apps/dashboard/src/components/workflow-editor/steps/chat/chat-tabs.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react';
 import { useFormContext } from 'react-hook-form';
 import { ChatEditor } from '@/components/workflow-editor/steps/chat/chat-editor';
 import { ChatEditorPreview } from '@/components/workflow-editor/steps/chat/chat-editor-preview';
-import { StepEditorProps } from '@/components/workflow-editor/steps/configure-step-template-form';
 import { CustomStepControls } from '@/components/workflow-editor/steps/controls/custom-step-controls';
+import { StepEditorProps } from '@/components/workflow-editor/steps/step-editor-types';
 import { TemplateTabs } from '@/components/workflow-editor/steps/template-tabs';
 import { useEditorPreview } from '../use-editor-preview';
