+++ apps/dashboard/src/components/workflow-editor/steps/sms/sms-tabs.tsx
@@ -1,10 +1,10 @@
 import { ResourceOriginEnum } from '@novu/shared';
 import { useState } from 'react';
 import { useFormContext } from 'react-hook-form';
-import { StepEditorProps } from '@/components/workflow-editor/steps/configure-step-template-form';
 import { CustomStepControls } from '@/components/workflow-editor/steps/controls/custom-step-controls';
 import { SmsEditor } from '@/components/workflow-editor/steps/sms/sms-editor';
 import { SmsEditorPreview } from '@/components/workflow-editor/steps/sms/sms-editor-preview';
+import { StepEditorProps } from '@/components/workflow-editor/steps/step-editor-types';
 import { TemplateTabs } from '@/components/workflow-editor/steps/template-tabs';
 import { useEditorPreview } from '../use-editor-preview';
