+++ apps/dashboard/src/components/workflow-editor/steps/step-editor-types.ts
@@ -1,0 +1,6 @@
+import { type StepResponseDto, type WorkflowResponseDto } from '@novu/shared';
+
+export type StepEditorProps = {
+  workflow: WorkflowResponseDto;
+  step: StepResponseDto;
+};
